Use Kotlin Flow & StateFlow for communication between view & presenter

+++ core/gamedex-core-filter/src/main/kotlin/com/gitlab/ykrasik/gamedex/core/filter/presenter/EditFilterPresenter.kt
@@ -22,8 +22,10 @@ import com.gitlab.ykrasik.gamedex.core.Presenter
 import com.gitlab.ykrasik.gamedex.core.ViewSession
 import com.gitlab.ykrasik.gamedex.core.filter.FilterService
 import com.gitlab.ykrasik.gamedex.core.task.TaskService
-import com.gitlab.ykrasik.gamedex.util.Try
+import com.gitlab.ykrasik.gamedex.util.IsValid
 import com.gitlab.ykrasik.gamedex.util.and
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -40,34 +42,31 @@ class EditFilterPresenter @Inject constructor(
 ) : Presenter<EditFilterView> {
     override fun present(view: EditFilterView) = object : ViewSession() {
         init {
-            view.initialNamedFilter.forEach { initialNamedFilter ->
-                view.name *= initialNamedFilter.id
-                view.filter *= initialNamedFilter.filter
-                view.isTag *= if (!initialNamedFilter.isAnonymous) initialNamedFilter.isTag else true
-            }
-            view.name.mapTo(view.nameIsValid) { name ->
-                Try { check(name.isNotBlank()) { "Name is required!" } }
-            }
+            view.name *= view.initialNamedFilter.onlyChangesFromView().map { it.id } withDebugName "name"
+            view.filter *= view.initialNamedFilter.onlyChangesFromView().map { it.filter } withDebugName "filter"
+            view.isTag *= view.initialNamedFilter.onlyChangesFromView().map { if (!it.isAnonymous) it.isTag else true } withDebugName "isTag"
 
-            view.name.combineLatest(view.nameIsValid, view.filter, view.filterIsValid, view.isTag) { name, nameIsValid, filter, filterIsValid, isTag ->
-                view.canAccept *= nameIsValid and filterIsValid and Try {
-                    val initialNamedFilter = view.initialNamedFilter.value
+            view.nameIsValid *= view.name.allValues().map { name ->
+                IsValid { check(name.isNotBlank()) { "Name is required!" } }
+            } withDebugName "nameIsValid"
+
+            view.canAccept *= combine(
+                view.nameIsValid,
+                view.filterValidatedValue.onlyChangesFromView(),
+                view.isTag.allValues()
+            ) { nameIsValid, filterValidatedValue, isTag ->
+                nameIsValid and filterValidatedValue.isValid and IsValid {
+                    val filter = filterValidatedValue.value
                     check(!filter.isEmpty) { "Filter is empty!" }
-                    if (name == initialNamedFilter.id &&
-                        filter.isEqual(initialNamedFilter.filter) &&
-                        isTag == initialNamedFilter.isTag/* &&
-                            view.excludedGames.map { it.id } == view.report?.excludedGames*/) {
-                        error("Nothing changed!")
-                    }
                     if (isTag && filter.find(Filter.FilterTag::class) != null) {
                         error("Filters that tag games may not depend on FilterTag filters!")
                     }
                 }
-            }
+            } withDebugName "canAccept"
 
 //            view.unexcludeGameActions.forEach { onUnexcludeGame(it) }
-            view.acceptActions.forEach { onAccept() }
-            view.cancelActions.forEach { onCancel() }
+            view.acceptActions.forEach(debugName = "onAccept") { onAccept() }
+            view.cancelActions.forEach(debugName = "onCancel") { onCancel() }
         }
 
 //        private fun onUnexcludeGame(game: Game) {
@@ -78,29 +77,28 @@ class EditFilterPresenter @Inject constructor(
         private suspend fun onAccept() {
             view.canAccept.assert()
 
-            val name = view.name.value
-            val initialNamedFilter = view.initialNamedFilter.value
-
-            val isRenamed = name != initialNamedFilter.id
-            val filterToOverwrite = if (isRenamed) {
-                filterService.userFilters.find { it.id == name }
-            } else {
-                null
-            }
-            if (filterToOverwrite != null && !view.confirmOverwrite(filterToOverwrite)) {
-                // There already is a filter with such a name which is not the filter we are editing, confirm user wants to overwrite.
-                return
-            }
-
-            val namedFilter = NamedFilter(
-                id = name,
-                filter = view.filter.value,
-                isTag = view.isTag.value
+            val initialNamedFilter = view.initialNamedFilter.v
+            val newNamedFilter = NamedFilter(
+                id = view.name.v,
+                filter = view.filter.v,
+                isTag = view.isTag.v
             )
+            if (newNamedFilter != initialNamedFilter.copy(timestamp = newNamedFilter.timestamp)) {
+                val isRenamed = newNamedFilter.id != initialNamedFilter.id
+                val filterToOverwrite = if (isRenamed) {
+                    filterService.userFilters.find { it.id == newNamedFilter.id }
+                } else {
+                    null
+                }
+                if (filterToOverwrite != null && !view.confirmOverwrite(filterToOverwrite)) {
+                    // There already is a filter with such a name which is not the filter we are editing, confirm user wants to overwrite.
+                    return
+                }
 
-            taskService.execute(filterService.save(namedFilter))
-            if (!initialNamedFilter.isAnonymous && isRenamed) {
-                taskService.execute(filterService.delete(initialNamedFilter))
+                taskService.execute(filterService.save(newNamedFilter))
+                if (!initialNamedFilter.isAnonymous && isRenamed) {
+                    taskService.execute(filterService.delete(initialNamedFilter))
+                }
             }
 
             hideView()
